Modified temp directory handling for some testers.
Some CPANTESTER boxes are not allowing writes to their /tmp directory
despite the fact that File::Spec->tmpdir is returning that directory.
This causes a large number of test failures.

Despite the fact that I don't like the practice, I've modified the tests
to work in a temporary directory in the project directory. Since we have
to be able to write here in order to build, that should be safe.