Add support for reconstituting objects from disk.
This code addresses a security hole relating to objects recreated from a
serialized format on disk. A malicious user could modify the disk image
a get a different object type loaded. This code only rebuilds the
objects if they are Tasks and discards all others.

Update copyrights and version numbers.